added the history feature
